+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/statistics/ui/stateselection/FederalStateAdapter.kt
@@ -1,16 +1,16 @@
-package de.rki.coronawarnapp.statistics.ui
+package de.rki.coronawarnapp.statistics.ui.stateselection
 
 import android.view.ViewGroup
 import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.databinding.FederalStateInputAdapterItemBinding
 import de.rki.coronawarnapp.ui.lists.BaseAdapter
 
-class FederalStateAdapter constructor(val onItemClickListener: (UserInfoItem) -> Unit) :
+class FederalStateAdapter constructor(val onItemClickListener: (ListItem) -> Unit) :
     BaseAdapter<FederalStateAdapter.VH>() {
 
-    private val internalData = mutableListOf<UserInfoItem>()
+    private val internalData = mutableListOf<ListItem>()
 
-    var data: List<UserInfoItem>
+    var data: List<ListItem>
         get() = internalData.toList()
         set(value) {
             internalData.clear()
@@ -34,7 +34,7 @@ class FederalStateAdapter constructor(val onItemClickListener: (UserInfoItem) ->
         private val viewBinding: FederalStateInputAdapterItemBinding =
             FederalStateInputAdapterItemBinding.bind(itemView)
 
-        fun bind(item: UserInfoItem) = viewBinding.apply {
+        fun bind(item: ListItem) = viewBinding.apply {
             label.text = item.label.get(context)
         }
     }
