+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/statistics/ui/stateselection/ListItem.kt
@@ -1,8 +1,8 @@
-package de.rki.coronawarnapp.statistics.ui
+package de.rki.coronawarnapp.statistics.ui.stateselection
 
 import de.rki.coronawarnapp.util.ui.LazyString
 
-data class UserInfoItem(
+data class ListItem(
     val label: LazyString,
     val data: Any
 )
