+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/main/MainActivityModule.kt
@@ -8,7 +8,7 @@ import de.rki.coronawarnapp.covidcertificate.DigitalCovidCertificateUIModule
 import de.rki.coronawarnapp.datadonation.analytics.ui.AnalyticsUIModule
 import de.rki.coronawarnapp.release.NewReleaseInfoFragment
 import de.rki.coronawarnapp.release.NewReleaseInfoFragmentModule
-import de.rki.coronawarnapp.statistics.ui.FederalStateSelectionModule
+import de.rki.coronawarnapp.statistics.ui.stateselection.FederalStateSelectionModule
 import de.rki.coronawarnapp.tracing.ui.details.TracingDetailsFragmentModule
 import de.rki.coronawarnapp.ui.coronatest.rat.profile.RATProfileUIModule
 import de.rki.coronawarnapp.ui.information.InformationFragmentModule
