+++ Corona-Warn-App/src/main/res/layout/federal_state_selection_fragment.xml
@@ -18,6 +18,7 @@
         tools:title="Ihr Alter/Bundesland/Kreis " />
 
     <androidx.recyclerview.widget.RecyclerView
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         android:id="@+id/input_list"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
